requirements.txt
fastapi
uvicorn
opencv-python
numpy
qdrant-client
python-multipart 
python-dotenv